reorganizing serializers into a separate folder
